ulong cleanup

+++ server/drizzled.cc
@@ -445,7 +445,7 @@ pthread_cond_t  COND_server_started;
 
 /* replication parameters, if master_host is not NULL, we are a slave */
 uint report_port= MYSQL_PORT;
-ulong master_retry_count=0;
+uint32_t master_retry_count= 0;
 char *master_info_file;
 char *relay_log_info_file, *report_user, *report_password, *report_host;
 char *opt_relay_logname = 0, *opt_relaylog_index_name=0;
